+++ src/onegov.election_day/onegov/election_day/forms/vote.py
@@ -22,24 +22,11 @@ class VoteForm(Form):
         default='simple'
     )
 
-    vote_de = StringField(
-        label=_("Vote (German)"),
+    domain = RadioField(
+        label=_("Type"),
         validators=[
             InputRequired()
         ]
-    )
-    vote_fr = StringField(
-        label=_("Vote (French)")
-    )
-    vote_it = StringField(
-        label=_("Vote (Italian)")
-    )
-    vote_rm = StringField(
-        label=_("Vote (Romansh)")
-    )
-
-    shortcode = StringField(
-        label=_("Shortcode")
     )
 
     date = DateField(
@@ -48,16 +35,101 @@ class VoteForm(Form):
         default=date.today
     )
 
-    domain = RadioField(
-        label=_("Type"),
-        validators=[
-            InputRequired()
-        ]
+    shortcode = StringField(
+        label=_("Shortcode")
+    )
+
+    vote_de = StringField(
+        label=_("German"),
+        fieldset=_("Title of the the vote/proposal"),
+        render_kw={'lang': 'de'}
+    )
+    vote_fr = StringField(
+        label=_("French"),
+        fieldset=_("Title of the the vote/proposal"),
+        render_kw={'lang': 'fr'}
+    )
+    vote_it = StringField(
+        label=_("Italian"),
+        fieldset=_("Title of the the vote/proposal"),
+        render_kw={'lang': 'it'}
+    )
+    vote_rm = StringField(
+        label=_("Romansh"),
+        fieldset=_("Title of the the vote/proposal"),
+        render_kw={'lang': 'rm'}
+    )
+
+    counter_proposal_de = StringField(
+        label=_("German"),
+        fieldset=_("Title of the counter proposal"),
+        render_kw={'lang': 'de'},
+        depends_on=('vote_type', 'complex')
+    )
+    counter_proposal_fr = StringField(
+        label=_("French"),
+        fieldset=_("Title of the counter proposal"),
+        render_kw={'lang': 'fr'},
+        depends_on=('vote_type', 'complex')
+    )
+    counter_proposal_it = StringField(
+        label=_("Italian"),
+        fieldset=_("Title of the counter proposal"),
+        render_kw={'lang': 'it'},
+        depends_on=('vote_type', 'complex')
+    )
+    counter_proposal_rm = StringField(
+        label=_("Romansh"),
+        fieldset=_("Title of the counter proposal"),
+        render_kw={'lang': 'rm'},
+        depends_on=('vote_type', 'complex')
+    )
+
+    tie_breaker_de = StringField(
+        label=_("German"),
+        fieldset=_("Title of the tie breaker"),
+        render_kw={'lang': 'de'},
+        depends_on=('vote_type', 'complex')
+    )
+    tie_breaker_fr = StringField(
+        label=_("French"),
+        fieldset=_("Title of the tie breaker"),
+        render_kw={'lang': 'fr'},
+        depends_on=('vote_type', 'complex'),
+    )
+    tie_breaker_it = StringField(
+        label=_("Italian"),
+        fieldset=_("Title of the tie breaker"),
+        render_kw={'lang': 'it'},
+        depends_on=('vote_type', 'complex')
+    )
+    tie_breaker_rm = StringField(
+        label=_("Romansh"),
+        fieldset=_("Title of the tie breaker"),
+        render_kw={'lang': 'rm'},
+        depends_on=('vote_type', 'complex')
     )
 
     related_link = URLField(
         label=_("Related link")
     )
+
+    def validate(self):
+        result = super(VoteForm, self).validate()
+        if not any((
+            self.vote_de.data,
+            self.vote_fr.data,
+            self.vote_it.data,
+            self.vote_rm.data
+        )):
+            message = _("Provide at least one title.")
+            self.vote_de.errors.append(message)
+            self.vote_fr.errors.append(message)
+            self.vote_it.errors.append(message)
+            self.vote_rm.errors.append(message)
+            result = False
+
+        return result
 
     def set_domain(self, principal):
         self.domain.choices = [
@@ -71,30 +143,76 @@ class VoteForm(Form):
         model.shortcode = self.shortcode.data
         model.related_link = self.related_link.data
 
-        model.title_translations = {}
-        model.title_translations['de_CH'] = self.vote_de.data
+        titles = {}
+        if self.vote_de.data:
+            titles['de_CH'] = self.vote_de.data
         if self.vote_fr.data:
-            model.title_translations['fr_CH'] = self.vote_fr.data
+            titles['fr_CH'] = self.vote_fr.data
         if self.vote_it.data:
-            model.title_translations['it_CH'] = self.vote_it.data
+            titles['it_CH'] = self.vote_it.data
         if self.vote_rm.data:
-            model.title_translations['rm_CH'] = self.vote_rm.data
+            titles['rm_CH'] = self.vote_rm.data
+        model.title_translations = titles
+
+        if model.type == 'complex':
+            titles = {}
+            if self.counter_proposal_de.data:
+                titles['de_CH'] = self.counter_proposal_de.data
+            if self.counter_proposal_fr.data:
+                titles['fr_CH'] = self.counter_proposal_fr.data
+            if self.counter_proposal_it.data:
+                titles['it_CH'] = self.counter_proposal_it.data
+            if self.counter_proposal_rm.data:
+                titles['rm_CH'] = self.counter_proposal_rm.data
+            model.counter_proposal.title_translations = titles
+
+            titles = {}
+            if self.tie_breaker_de.data:
+                titles['de_CH'] = self.tie_breaker_de.data
+            if self.tie_breaker_fr.data:
+                titles['fr_CH'] = self.tie_breaker_fr.data
+            if self.tie_breaker_it.data:
+                titles['it_CH'] = self.tie_breaker_it.data
+            if self.tie_breaker_rm.data:
+                titles['rm_CH'] = self.tie_breaker_rm.data
+            model.tie_breaker.title_translations = titles
 
     def apply_model(self, model):
-        self.vote_de.data = model.title_translations['de_CH']
-        self.vote_fr.data = model.title_translations.get('fr_CH')
-        self.vote_it.data = model.title_translations.get('it_CH')
-        self.vote_rm.data = model.title_translations.get('rm_CH')
+        titles = model.title_translations or {}
+        self.vote_de.data = titles.get('de_CH')
+        self.vote_fr.data = titles.get('fr_CH')
+        self.vote_it.data = titles.get('it_CH')
+        self.vote_rm.data = titles.get('rm_CH')
 
         self.date.data = model.date
         self.domain.data = model.domain
         self.shortcode.data = model.shortcode
         self.related_link.data = model.related_link
+
         if model.type == 'complex':
             self.vote_type.choices = [
                 ('complex', _("Vote with Counter-Proposal"))
             ]
             self.vote_type.data = 'complex'
+
+            titles = model.counter_proposal.title_translations or {}
+            self.counter_proposal_de.data = titles.get('de_CH')
+            self.counter_proposal_fr.data = titles.get('fr_CH')
+            self.counter_proposal_it.data = titles.get('it_CH')
+            self.counter_proposal_rm.data = titles.get('rm_CH')
+
+            titles = model.tie_breaker.title_translations or {}
+            self.tie_breaker_de.data = titles.get('de_CH')
+            self.tie_breaker_fr.data = titles.get('fr_CH')
+            self.tie_breaker_it.data = titles.get('it_CH')
+            self.tie_breaker_rm.data = titles.get('rm_CH')
+
         else:
             self.vote_type.choices = [('simple', _("Simple Vote"))]
             self.vote_type.data = 'simple'
+
+            for fieldset in self.fieldsets:
+                if fieldset.label == 'Title of the counter proposal':
+                    fieldset.label = None
+                if fieldset.label == 'Title of the tie breaker':
+                    fieldset.label = None