+++ src/onegov.election_day/onegov/election_day/tests/test_forms.py
@@ -89,11 +89,12 @@ def test_vote_form_domains():
 
 def test_vote_form_model(election_day_app):
     model = Vote()
-    model.title = 'Vote (DE)'
-    model.title_translations['de_CH'] = 'Vote (DE)'
-    model.title_translations['fr_CH'] = 'Vote (FR)'
-    model.title_translations['it_CH'] = 'Vote (IT)'
-    model.title_translations['rm_CH'] = 'Vote (RM)'
+    model.title_translations = {
+        'de_CH': 'Vote (DE)',
+        'fr_CH': 'Vote (FR)',
+        'it_CH': 'Vote (IT)',
+        'rm_CH': 'Vote (RM)',
+    }
     model.date = date.today()
     model.domain = 'federation'
     model.shortcode = 'xy'
@@ -106,11 +107,23 @@ def test_vote_form_model(election_day_app):
     assert form.vote_fr.data == 'Vote (FR)'
     assert form.vote_it.data == 'Vote (IT)'
     assert form.vote_rm.data == 'Vote (RM)'
+    assert form.counter_proposal_de.data is None
+    assert form.counter_proposal_fr.data is None
+    assert form.counter_proposal_it.data is None
+    assert form.counter_proposal_rm.data is None
+    assert form.tie_breaker_de.data is None
+    assert form.tie_breaker_fr.data is None
+    assert form.tie_breaker_it.data is None
+    assert form.tie_breaker_rm.data is None
     assert form.date.data == date.today()
     assert form.domain.data == 'federation'
     assert form.shortcode.data == 'xy'
     assert form.related_link.data == 'http://u.rl'
     assert form.vote_type.data == 'simple'
+
+    fieldsets = [f.label for f in form.fieldsets if f.label]
+    assert 'Title of the counter proposal' not in fieldsets
+    assert 'Title of the tie breaker' not in fieldsets
 
     form.vote_de.data = 'A Vote (DE)'
     form.vote_fr.data = 'A Vote (FR)'
@@ -124,22 +137,112 @@ def test_vote_form_model(election_day_app):
 
     form.update_model(model)
 
-    assert model.title == 'A Vote (DE)'
-    assert model.title_translations['de_CH'] == 'A Vote (DE)'
-    assert model.title_translations['fr_CH'] == 'A Vote (FR)'
-    assert model.title_translations['it_CH'] == 'A Vote (IT)'
-    assert model.title_translations['rm_CH'] == 'A Vote (RM)'
+    assert model.title_translations == {
+        'de_CH': 'A Vote (DE)',
+        'fr_CH': 'A Vote (FR)',
+        'it_CH': 'A Vote (IT)',
+        'rm_CH': 'A Vote (RM)',
+    }
     assert model.date == date(2016, 1, 1)
     assert model.domain == 'canton'
     assert model.shortcode == 'yz'
     assert model.related_link == 'http://ur.l'
     assert model.type == 'simple'
 
-    model = ComplexVote(title='Vote')
 
+def test_vote_form_model_complex(election_day_app):
+    model = ComplexVote()
+    model.title_translations = {
+        'de_CH': 'Vote (DE)',
+        'fr_CH': 'Vote (FR)',
+        'it_CH': 'Vote (IT)',
+        'rm_CH': 'Vote (RM)',
+    }
+    model.counter_proposal.title_translations = {
+        'de_CH': 'Counter Proposal (DE)',
+        'fr_CH': 'Counter Proposal (FR)',
+        'it_CH': 'Counter Proposal (IT)',
+        'rm_CH': 'Counter Proposal (RM)',
+    }
+    model.tie_breaker.title_translations = {
+        'de_CH': 'Tie Breaker (DE)',
+        'fr_CH': 'Tie Breaker (FR)',
+        'it_CH': 'Tie Breaker (IT)',
+        'rm_CH': 'Tie Breaker (RM)',
+    }
+    model.date = date.today()
+    model.domain = 'federation'
+    model.shortcode = 'xy'
+    model.related_link = 'http://u.rl'
+
+    form = VoteForm()
     form.apply_model(model)
 
+    assert form.vote_de.data == 'Vote (DE)'
+    assert form.vote_fr.data == 'Vote (FR)'
+    assert form.vote_it.data == 'Vote (IT)'
+    assert form.vote_rm.data == 'Vote (RM)'
+    assert form.counter_proposal_de.data == 'Counter Proposal (DE)'
+    assert form.counter_proposal_fr.data == 'Counter Proposal (FR)'
+    assert form.counter_proposal_it.data == 'Counter Proposal (IT)'
+    assert form.counter_proposal_rm.data == 'Counter Proposal (RM)'
+    assert form.tie_breaker_de.data == 'Tie Breaker (DE)'
+    assert form.tie_breaker_fr.data == 'Tie Breaker (FR)'
+    assert form.tie_breaker_it.data == 'Tie Breaker (IT)'
+    assert form.tie_breaker_rm.data == 'Tie Breaker (RM)'
+    assert form.date.data == date.today()
+    assert form.domain.data == 'federation'
+    assert form.shortcode.data == 'xy'
+    assert form.related_link.data == 'http://u.rl'
     assert form.vote_type.data == 'complex'
+
+    fieldsets = [f.label for f in form.fieldsets if f.label]
+    assert 'Title of the counter proposal' in fieldsets
+    assert 'Title of the tie breaker' in fieldsets
+
+    form.vote_de.data = 'A Vote (DE)'
+    form.vote_fr.data = 'A Vote (FR)'
+    form.vote_it.data = 'A Vote (IT)'
+    form.vote_rm.data = 'A Vote (RM)'
+    form.counter_proposal_de.data = 'The Counter Proposal (DE)'
+    form.counter_proposal_fr.data = 'The Counter Proposal (FR)'
+    form.counter_proposal_it.data = 'The Counter Proposal (IT)'
+    form.counter_proposal_rm.data = 'The Counter Proposal (RM)'
+    form.tie_breaker_de.data = 'The Tie Breaker (DE)'
+    form.tie_breaker_fr.data = 'The Tie Breaker (FR)'
+    form.tie_breaker_it.data = 'The Tie Breaker (IT)'
+    form.tie_breaker_rm.data = 'The Tie Breaker (RM)'
+    form.date.data = date(2016, 1, 1)
+    form.domain.data = 'canton'
+    form.shortcode.data = 'yz'
+    form.related_link.data = 'http://ur.l'
+    form.vote_type.data = 'complex'
+
+    form.update_model(model)
+
+    assert model.title_translations == {
+        'de_CH': 'A Vote (DE)',
+        'fr_CH': 'A Vote (FR)',
+        'it_CH': 'A Vote (IT)',
+        'rm_CH': 'A Vote (RM)',
+    }
+    assert model.counter_proposal.title_translations == {
+        'de_CH': 'The Counter Proposal (DE)',
+        'fr_CH': 'The Counter Proposal (FR)',
+        'it_CH': 'The Counter Proposal (IT)',
+        'rm_CH': 'The Counter Proposal (RM)',
+    }
+    assert model.tie_breaker.title_translations == {
+        'de_CH': 'The Tie Breaker (DE)',
+        'fr_CH': 'The Tie Breaker (FR)',
+        'it_CH': 'The Tie Breaker (IT)',
+        'rm_CH': 'The Tie Breaker (RM)',
+    }
+    assert model.date == date(2016, 1, 1)
+    assert model.domain == 'canton'
+    assert model.shortcode == 'yz'
+    assert model.related_link == 'http://ur.l'
+    assert model.type == 'complex'
 
 
 def test_election_form_domains():